refactoring names of files
