+++ client/store/index.js
@@ -2,9 +2,9 @@ import {createStore, combineReducers, applyMiddleware} from 'redux'
 import createLogger from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
-import user from './user'
-import products from './products'
-import cart from './cart'
+import user from './user_store'
+import products from './products_store'
+import cart from './cart_store'
 
 const reducer = combineReducers({user, products, cart})
 const middleware = composeWithDevTools(
@@ -13,4 +13,4 @@ const middleware = composeWithDevTools(
 const store = createStore(reducer, middleware)
 
 export default store
-export * from './user'
+export * from './user_store'
