+++ server/api/index.js
@@ -1,10 +1,10 @@
 const router = require('express').Router()
 module.exports = router
 
-router.use('/users', require('./users'))
-router.use('/products', require('./products'))
-router.use('/carts', require('./carts'))
-router.use('/cartProducts', require('./cartProducts'))
+router.use('/users', require('./users_route'))
+router.use('/products', require('./products_route'))
+router.use('/carts', require('./carts_route'))
+router.use('/cartProducts', require('./cartProducts_route'))
 
 router.use((req, res, next) => {
   const error = new Error('Not Found')
